make alert appear so if user submits textarea with less than 15 characters, also confirm if user wants to delete a review or not

+++ week5-Journal/Client/src/app.js
@@ -106,17 +106,15 @@
 
 // // postReviews();
 
-const form = document.getElementById("review_form");
-const commentContainer = document.getElementById("comment-container");
+const form = document.getElementById('review_form');
+const commentContainer = document.getElementById('comment-container');
 
-
-const BASE_URL = "https://week5-project-09nc.onrender.com";
+const BASE_URL = 'https://week5-project-09nc.onrender.com';
 
 const fetchReviews = async () => {
   const response = await fetch(`${BASE_URL}/reviews`);
   const data = await response.json();
-  console.log(data);
-  commentContainer.innerHTML = "";
+  commentContainer.innerHTML = '';
   data
     .sort((a, b) => new Date(b.created_at) - new Date(a.created_at))
     .forEach((review) => createReview(review));
@@ -126,14 +124,20 @@ fetchReviews();
 
 const postReviews = async (e) => {
   e.preventDefault();
+  const reviewDescription = document.querySelector('#review_description').value;
+
+  if (reviewDescription.length < 15) {
+    alert('Must be more than 15 characters long');
+    return;
+  }
 
   const formData = new FormData(form);
   const formEntries = Object.fromEntries(formData);
 
   const response = await fetch(`${BASE_URL}/reviews`, {
-    method: "POST",
+    method: 'POST',
     headers: {
-      "Content-Type": "application/json",
+      'Content-Type': 'application/json',
     },
     body: JSON.stringify(formEntries),
   });
@@ -145,19 +149,23 @@ const postReviews = async (e) => {
       console.log("couldn't fetch reviews");
     }
   } catch (error) {
-    console.log("Error:", error);
+    console.log('Error:', error);
   }
 };
 
-form.addEventListener("submit", postReviews);
+form.addEventListener('submit', postReviews);
 
 async function handleDeleteReview() {
+  const deleteConfirmation = confirm(
+    'Are you sure you want to delete this review?'
+  );
+
+  if (!deleteConfirmation) {
+    return;
+  }
+
   const commentDiv = this.closest('div');
   const id = commentDiv.dataset.id;
-  console.log(id);
-
-  console.log('first');
-
   try {
     const response = await fetch(`${BASE_URL}/reviews`, {
       method: 'DELETE',
@@ -176,24 +184,24 @@ async function handleDeleteReview() {
 }
 
 function createReview(review) {
-  const reviewDiv = document.createElement("div");
-  reviewDiv.className = "review";
+  const reviewDiv = document.createElement('div');
+  reviewDiv.className = 'review';
   reviewDiv.dataset.id = review.id;
 
-  const username = createParagraph("username", review.name);
+  const username = createParagraph('username', review.name);
 
   const description = createParagraph(
-    "review-description",
+    'review-description',
     review.review_description
   );
 
-  const image = createImage("review-image", review.image_url);
+  const image = createImage('review-image', review.image_url);
 
-  const formattedDate = new Intl.DateTimeFormat("en-GB").format(
+  const formattedDate = new Intl.DateTimeFormat('en-GB').format(
     new Date(review.created_at)
   );
 
-  const reviewDate = createParagraph("date", formattedDate);
+  const reviewDate = createParagraph('date', formattedDate);
 
   const rating = createParagraph('rating', `${'⭐'.repeat(review.rating)}`);
 
@@ -218,14 +226,14 @@ function createReview(review) {
 }
 
 function createParagraph(className, text) {
-  const p = document.createElement("p");
+  const p = document.createElement('p');
   p.className = className;
   p.textContent = text;
   return p;
 }
 
 function createImage(className, url) {
-  const image = document.createElement("img");
+  const image = document.createElement('img');
   image.className = className;
   image.src = url;
   return image;
@@ -233,7 +241,6 @@ function createImage(className, url) {
 
 function createButton(className, text) {
   const button = document.createElement('i');
-
   button.className = className;
   button.title = text;
   return button;
